+++ src/components/Visit/visitlist.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react'
 
-import { useFirestore } from '@/context/firestore'
+import { useFirestore } from '@/context/Firestore'
 import { SelectOption, VisitData } from '@/types/types'
 
 import VisitInstance from './visitinstance'
