+++ src/pages/_app.tsx
@@ -1,20 +1,18 @@
 import { AppProps } from 'next/app'
 
-import { AuthContextProvider } from '@/context/AuthContext'
-import FirestoreProvider from '@/context/firestore'
+import { AuthContextProvider } from '@/context/Auth'
+import FirestoreProvider from '@/context/Firestore'
 
 import '@/styles/main.css'
 
 const POOPS = ({ Component, pageProps }: AppProps) => {
   return (
-    <>
-      <AuthContextProvider>
-        <FirestoreProvider>
-          <Component {...pageProps} />
-          {/* <NavBar /> */}
-        </FirestoreProvider>
-      </AuthContextProvider>
-    </>
+    <AuthContextProvider>
+      <FirestoreProvider>
+        <Component {...pageProps} />
+        {/* <NavBar /> */}
+      </FirestoreProvider>
+    </AuthContextProvider>
   )
 }
