+++ src/pages/profile.tsx
@@ -2,7 +2,7 @@ import Header from '@/components/Header'
 import LoginPanel from '@/components/Login/LoginPanel'
 import NavBar from '@/components/NavBar'
 // import { withProtected } from '@/components/PrivateRoute'
-import { useAuth } from '@/context/AuthContext'
+import { useAuth } from '@/context/Auth'
 
 const Profile = () => {
   const { logOut } = useAuth()
